# web/static/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_alchemillahq_sylve_internal.APIResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.CPUPinning:
    properties:
      cores:
        items:
          type: integer
        minItems: 1
        type: array
      socket:
        minimum: 0
        type: integer
    required:
    - cores
    - socket
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.ModifyCPURequest:
    properties:
      cpuCores:
        type: integer
      cpuPinning:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.CPUPinning'
        type: array
      cpuSockets:
        type: integer
      cpuThreads:
        type: integer
    required:
    - cpuCores
    - cpuSockets
    - cpuThreads
    type: object
  github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.ModifyVNCRequest:
    properties:
      vncEnabled:
        type: boolean
      vncPassword:
        type: string
      vncPort:
        type: integer
      vncResolution:
        type: string
      vncWait:
        type: boolean
    required:
    - vncPassword
    - vncPort
    - vncResolution
    type: object
...
  internal_handlers_vm.ModifyPassthroughRequest:
    properties:
      pciDevices:
        items:
          type: integer
        type: array
    required:
    - pciDevices
    type: object
  internal_handlers_vm.ModifyRAMRequest:
    properties:
      ram:
        type: integer
    required:
    - ram
    type: object
...
host: sylve.lan:8181
info:
  contact:
    email: hello@alchemilla.io
    name: Alchemilla Ventures Pvt. Ltd.
    url: https://alchemilla.io
  description: Sylve is a lightweight GUI for managing Bhyve, Jails, ZFS, networking,
    and more on FreeBSD.
  license:
    name: BSD-2-Clause
    url: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  termsOfService: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  title: Sylve API
  version: 0.0.1
paths:
...
  /hardware/cpu/:rid:
    put:
      consumes:
      - application/json
      description: Modify the CPU configuration of a virtual machine
      parameters:
      - description: Modify CPU Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.ModifyCPURequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Modify CPU of a Virtual Machine
      tags:
      - VM
  /hardware/ppt/:rid:
    put:
      consumes:
      - application/json
      description: Modify the PCI Passthrough devices of a virtual machine
      parameters:
      - description: Modify PCI Devices Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_vm.ModifyPassthroughRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Modify PCI Devices of a Virtual Machine
      tags:
      - VM
  /hardware/ram/:rid:
    put:
      consumes:
      - application/json
      description: Modify the RAM configuration of a virtual machine
      parameters:
      - description: Modify RAM Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_vm.ModifyRAMRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Modify RAM of a Virtual Machine
      tags:
      - VM
  /hardware/vnc/:rid:
    put:
      consumes:
      - application/json
      description: Modify the VNC configuration of a virtual machine
      parameters:
      - description: Modify VNC Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_libvirt.ModifyVNCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Modify VNC of a Virtual Machine
      tags:
      - VM
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"